+++ testing/conftest.py
@@ -1,5 +1,5 @@
 import pytest
 
 pytest_plugins = [
-    "fixtures.start",
+    "fixtures.main",
 ]
