+++ testing/test_send_dynamic.py
@@ -12,6 +12,8 @@ import daf
 
 # SIMULATION
 #
+
+
 # CONFIGURATION
 TEST_TOKEN = os.environ.get("DISCORD_TOKEN")
 TEST_GUILD_ID = 863071397207212052
@@ -25,10 +27,6 @@ TEST_CHANNEL_NUM = 2
 async def test_text_message_send():
     "This tests if all the text messages succeed in their sends"
     text_channels = []
-    guild = daf.get_guild_user(TEST_GUILD_ID)
-    user = daf.get_guild_user(TEST_USER_ID)
-    text_message = None
-    direct_message = None
     try:
         @daf.data_function
         def dynamic_getter(items: list):
@@ -59,6 +57,10 @@ async def test_text_message_send():
                                                    start_in=timedelta(0), remove_after=None)
 
         # Initialize objects
+        guild = daf.GUILD(TEST_GUILD_ID)
+        user = daf.USER(TEST_USER_ID)
+        await guild.initialize()
+        await user.initialize()
         await guild.add_message(text_message)
         await user.add_message(direct_message)
 
@@ -90,12 +92,6 @@ async def test_text_message_send():
             with suppress(daf.discord.HTTPException):
                 await channel.delete()
         
-        with suppress(ValueError):
-            if text_message is not None:
-                guild.remove_message(text_message)
-        with suppress(ValueError):
-            if direct_message is not None:
-                user.remove_message(direct_message)
 
 
 
@@ -104,10 +100,7 @@ async def test_voice_message_send():
     "This tests if all the voice messages succeed in their sends"
     
     voice_channels = []
-    voice_message = None
-    guild = daf.get_guild_user(TEST_GUILD_ID)
     try:
-        await daf.initialize(token=TEST_TOKEN)
         client = daf.get_client()
         dc_guild = client.get_guild(TEST_GUILD_ID)
         dc_test_cat = client.get_channel(TEST_CAT_CHANNEL_ID)
@@ -126,10 +119,11 @@ async def test_voice_message_send():
         for i in range(1, TEST_CHANNEL_NUM + 1):
             voice_channels.append(await dc_guild.create_voice_channel(TEST_CHANNEL_FORMAT.format(i), category=dc_test_cat))
 
+        guild = daf.GUILD(TEST_GUILD_ID)
         voice_message = daf.message.VoiceMESSAGE(None, timedelta(seconds=20), VOICE_MESSAGE_TEST_MESSAGE, voice_channels,
                                                 volume=50, start_in=timedelta(), remove_after=None)
         
-
+        await guild.initialize()
         await guild.add_message(voice_message)
         for length, audio in data_:
             # Send
@@ -146,6 +140,5 @@ async def test_voice_message_send():
             with suppress(daf.discord.HTTPException):
                 await channel.delete()
 
-        with suppress(ValueError):
-            if voice_message is not None:
-                guild.remove_message(voice_message)
+
+
