+++ testing/test_send_normal.py
@@ -9,6 +9,7 @@ import pytest
 import daf
 
 
+
 # SIMULATION
 #
 
@@ -27,14 +28,9 @@ TEST_CHANNEL_NUM = 2
 async def test_text_message_send():
     "This tests if all the text messages succeed in their sends"
     text_channels = []
-    guild = daf.get_guild_user(TEST_GUILD_ID)
-    user = daf.get_guild_user(TEST_USER_ID)
-    text_message = None
-    direct_message = None
     try:
         TEXT_MESSAGE_TEST_MESSAGE = "Hello world", daf.discord.Embed(title="Hello world")
 
-        await daf.initialize(token=TEST_TOKEN)
         client = daf.get_client()
         dc_guild = client.get_guild(TEST_GUILD_ID)
         dc_test_cat = client.get_channel(TEST_CAT_CHANNEL_ID)
@@ -44,14 +40,15 @@ async def test_text_message_send():
             text_channels.append(await dc_guild.create_text_channel(TEST_CHANNEL_FORMAT.format(i), category=dc_test_cat))
 
 
-        # Create MESSAGE objects
+        guild = daf.GUILD(TEST_GUILD_ID)
+        user = daf.USER(TEST_USER_ID)
         text_message = daf.message.TextMESSAGE(None, timedelta(seconds=5), TEXT_MESSAGE_TEST_MESSAGE, text_channels,
                                             "send", start_in=timedelta(), remove_after=None)
-
         direct_message = daf.message.DirectMESSAGE(None, timedelta(seconds=5), TEXT_MESSAGE_TEST_MESSAGE, "send",
                                                    start_in=timedelta(0), remove_after=None)
-
         # Initialize objects
+        await guild.initialize()
+        await user.initialize()
         await guild.add_message(text_message)
         await user.add_message(direct_message)
 
@@ -87,24 +84,14 @@ async def test_text_message_send():
             with suppress(daf.discord.HTTPException):
                 await channel.delete()
 
-        with suppress(ValueError):
-            if text_message is not None:
-                guild.remove_message(text_message)
-        with suppress(ValueError):
-            if direct_message is not None:
-                user.remove_message(direct_message)
-
 
 @pytest.mark.asyncio
 async def test_voice_message_send():
     "This tests if all the voice messages succeed in their sends"
     voice_channels = []
-    guild = daf.get_guild_user(TEST_GUILD_ID)
-    voice_message = None
     try:
         VOICE_MESSAGE_TEST_MESSAGE = daf.AUDIO("https://www.youtube.com/watch?v=tCDvOQI3pco") # 10 second countdown
 
-        await daf.initialize(token=TEST_TOKEN)
         client = daf.get_client()
         dc_guild = client.get_guild(TEST_GUILD_ID)
         dc_test_cat = client.get_channel(TEST_CAT_CHANNEL_ID)
@@ -113,11 +100,11 @@ async def test_voice_message_send():
         for i in range(1, TEST_CHANNEL_NUM + 1):
             voice_channels.append(await dc_guild.create_voice_channel(TEST_CHANNEL_FORMAT.format(i), category=dc_test_cat))
 
+        guild = daf.GUILD(TEST_GUILD_ID)
         voice_message = daf.message.VoiceMESSAGE(None, timedelta(seconds=20), VOICE_MESSAGE_TEST_MESSAGE, voice_channels,
                                                 volume=50, start_in=timedelta(), remove_after=None)
         
         # Initialize objects
-        guild = daf.GUILD(TEST_GUILD_ID)
         await guild.initialize()
         await guild.add_message(voice_message)
 
@@ -135,7 +122,3 @@ async def test_voice_message_send():
         for channel in voice_channels:
             with suppress(daf.discord.HTTPException):
                 await channel.delete()
-        
-        with suppress(ValueError):
-            if voice_message is not None:
-                guild.remove_message(voice_message)
